+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Add 'manualDisabled' to heater settings (parallel to AC-Devices)
+
 ## 2.19.0 (2023-10-21)
 
 * (theimo1221) Enhance stability of ioBroker Devices to correctly have a room.